interviewGPT-be/requirements.txt
PyPDF2
python-dotenv
openai
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
flask-sqlalchemy==3.0.5
Werkzeug==2.3.6


